Revert "i made a booboo"
This reverts commit bac8633be7a3c15e92f86def1158c1168a19a48a.

+++ src/MAndApps/apps/spacewars/entity/Player.java
@@ -21,7 +21,7 @@ public class Player extends Entity {
 	private boolean goBoom = false;
 
 	public Player() {
-		super(512, 550, WIDTH, HEIGHT);
+		super((int) 512, (int) 550, WIDTH, HEIGHT);
 	}
 
 	@Override
