+++ src/MAndApps/apps/spacewars/entity/enemy/BlueEnemy.java
@@ -137,7 +137,7 @@ public class BlueEnemy extends Enemy {
 		g.setColor(color);
 		int temp;
 		try{
-			temp = r.nextInt((int)(0-(absoluteTime/20d))+5);
+			temp = r.nextInt((int)(0-((double)absoluteTime/20d))+5);
 		}catch(Exception e){
 			temp = 0;
 		}
